himiao-backend/requirements.txt
# ── Web 框架 ──
fastapi==0.115.0
uvicorn[standard]==0.32.0

# ── 数据库 ──
sqlalchemy==2.0.36
# 上云时加（PostgreSQL 驱动，无需修改 ORM 代码）：
# psycopg2-binary==2.9.9

# ── 配置管理 ──
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
email-validator==2.2.0
bcrypt==4.0.1

# ── 鉴权 ──
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── 邮件（Phase 3）──
# resend==2.4.0

# ── Phase 3：爬虫 + OCR（按需在 scraper 容器中安装）──
# scrapy==2.11.2
# playwright==1.48.0
# paddleocr==2.7.3
# pdfplumber==0.11.4
# pandas==2.2.3
httpx==0.28.1
beautifulsoup4==4.12.3
pdfplumber==0.11.4
pypdf==5.1.0
python-multipart==0.0.20   # File/Form upload 必须
